+++ website/react-magma-docs/src/components/SlidingDrawer/index.js
@@ -3,7 +3,6 @@ import { css, keyframes } from '@emotion/core';
 import styled from '@emotion/styled';
 import FocusLock from 'react-focus-lock';
 import { Logo } from '../Logo';
-import { LogoLink } from '../Masthead';
 import { MenuIcon, CloseIcon } from 'react-magma-icons';
 import { Container, IconButton, magma, Spacer } from 'react-magma-dom';
 import { MainNav } from '../MainNav';
@@ -137,13 +136,8 @@ export class SlidingDrawer extends React.Component {
     const MenuButton = styled.span`
       position: fixed;
       top: 4px;
-      left: 2px;
+      left: 6px;
       z-index: 11;
-
-      @media (min-width: 769px) {
-        top: 20px;
-      }
-
       @media (min-width: 1024px) {
         display: none;
       }
@@ -190,8 +184,8 @@ export class SlidingDrawer extends React.Component {
                 <IconButton
                   aria-label="Open navigation menu"
                   aria-expanded={isOpen}
+                  color="secondary"
                   icon={<MenuIcon />}
-                  isInverse
                   onClick={this.openMenu}
                   ref={this.toggleButtonRef}
                   variant="link"